+++ src/public/GetMTProjectInfo.ps1
@@ -43,6 +43,7 @@ function Get-MTProjectInfo {
     $Out['PrivateDir'] = [System.IO.Path]::Join($ProjectRoot, 'src', 'private')
     $Out['ClassesDir'] = [System.IO.Path]::Join($ProjectRoot, 'src', 'classes')
     $Out['ResourcesDir'] = [System.IO.Path]::Join($ProjectRoot, 'src', 'resources')
+    $Out['TestsDir'] = [System.IO.Path]::Join($ProjectRoot, 'tests')
     $Out['DocsDir'] = [System.IO.Path]::Join($ProjectRoot, 'docs')
     $Out['OutputDir'] = [System.IO.Path]::Join($ProjectRoot, 'dist')  
     $Out['OutputModuleDir'] = [System.IO.Path]::Join($Out.OutputDir, $ProjectName)